Rename 'api' folder to 'lib'
