--- apps/nextjs-website/src/api/types/page.ts
@@ -1,3 +1,0 @@
-import { Path } from '@/api/types/path';
-
-export type Page<T> = T & Path;
