+++ apps/nextjs-website/src/features/appIO/appIO.ts
@@ -1,4 +1,4 @@
-import { Product } from '@/api/types/product';
+import { Product } from '@/lib/types/product';
 import { appIoOverviewPath } from '@/features/appIO/overviewPath';
 
 export const appIO: Product = {
