+++ apps/nextjs-website/src/features/appIO/overviewPath.ts
@@ -1,4 +1,4 @@
-import { Path } from '@/api/types/path';
+import { Path } from '@/lib/types/path';
 
 export const appIoOverviewPath: Path = {
   name: 'App IO overview',
