+++ apps/nextjs-website/src/lib/types/overviewData.ts
@@ -1,4 +1,4 @@
-import { Page } from '@/api/types/page';
+import { Page } from '@/lib/types/page';
 import { OverviewPageProps } from '@/pages/[productSlug]/overview';
 
 export type OverviewData = Page<OverviewPageProps>;
