+++ apps/nextjs-website/src/lib/types/page.ts
@@ -1,0 +1,3 @@
+import { Path } from '@/lib/types/path';
+
+export type Page<T> = T & Path;
