+++ apps/nextjs-website/src/lib/types/path.ts
